Move main extension files, rename secondary extension files
