+++ lib/pages/chat/chat_view.dart
@@ -9,7 +9,7 @@ import 'package:fluffychat/pages/chat/reactions_picker.dart';
 import 'package:fluffychat/pages/chat/reply_display.dart';
 import 'package:fluffychat/pangea/choreographer/widgets/has_error_button.dart';
 import 'package:fluffychat/pangea/choreographer/widgets/language_permissions_warning_buttons.dart';
-import 'package:fluffychat/pangea/extensions/pangea_room_extension.dart';
+import 'package:fluffychat/pangea/extensions/pangea_room_extension/pangea_room_extension.dart';
 import 'package:fluffychat/pangea/pages/class_analytics/measure_able.dart';
 import 'package:fluffychat/utils/account_config.dart';
 import 'package:fluffychat/widgets/chat_settings_popup_menu.dart';
@@ -116,8 +116,10 @@ class ChatView extends StatelessWidget {
       // #Pangea
     } else {
       return [
-        ChatSettingsPopupMenu(controller.room,
-            (!controller.room.isDirectChat && !controller.room.isArchived)),
+        ChatSettingsPopupMenu(
+          controller.room,
+          (!controller.room.isDirectChat && !controller.room.isArchived),
+        ),
       ];
     }
