+++ lib/pangea/controllers/space_rules_edit_controller.dart
@@ -1,7 +1,7 @@
+import 'package:fluffychat/pangea/constants/pangea_event_types.dart';
 import 'package:matrix/matrix.dart';
 
-import 'package:fluffychat/pangea/constants/pangea_event_types.dart';
-import '../extensions/pangea_room_extension.dart';
+import '../extensions/pangea_room_extension/pangea_room_extension.dart';
 import '../models/class_model.dart';
 
 class RoomRulesEditController {
