+++ lib/pangea/extensions/pangea_room_extension/class_and_exchange_settings_extension.dart
@@ -1,4 +1,4 @@
-part of "../pangea_room_extension.dart";
+part of "pangea_room_extension.dart";
 
 extension PangeaRoom3 on Room {
   DateTime? get _rulesUpdatedAt {
