+++ lib/pangea/extensions/pangea_room_extension/room_settings_extension.dart
@@ -1,4 +1,4 @@
-part of "../pangea_room_extension.dart";
+part of "pangea_room_extension.dart";
 
 extension PangeaRoom6 on Room {
   PangeaRoomRules? get _pangeaRoomRules {
