+++ lib/pangea/pages/analytics/class_list/class_list_view.dart
@@ -1,4 +1,4 @@
-import 'package:fluffychat/pangea/extensions/client_extension.dart';
+import 'package:fluffychat/pangea/extensions/client_extension/client_extension.dart';
 import 'package:fluffychat/pangea/pages/analytics/analytics_list_tile.dart';
 import 'package:fluffychat/pangea/pages/analytics/time_span_menu_button.dart';
 import 'package:flutter/material.dart';
