+++ lib/pangea/pages/class_settings/class_name_header.dart
@@ -1,10 +1,8 @@
+import 'package:fluffychat/pages/chat_details/chat_details.dart';
+import 'package:fluffychat/pangea/extensions/pangea_room_extension/pangea_room_extension.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/visibility.dart' as visible;
-
 import 'package:matrix/matrix.dart';
-
-import 'package:fluffychat/pages/chat_details/chat_details.dart';
-import 'package:fluffychat/pangea/extensions/pangea_room_extension.dart';
 
 class ClassNameHeader extends StatelessWidget {
   final Room room;
