+++ lib/pangea/pages/class_settings/p_class_widgets/class_invitation_buttons.dart
@@ -1,16 +1,15 @@
+import 'package:fluffychat/pangea/config/environment.dart';
+import 'package:fluffychat/pangea/constants/url_query_parameter_keys.dart';
+import 'package:fluffychat/pangea/extensions/pangea_room_extension/pangea_room_extension.dart';
+import 'package:fluffychat/widgets/matrix.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:go_router/go_router.dart';
 import 'package:matrix/matrix.dart';
 import 'package:universal_html/html.dart' as html;
 
-import 'package:fluffychat/pangea/config/environment.dart';
-import 'package:fluffychat/pangea/constants/url_query_parameter_keys.dart';
-import 'package:fluffychat/pangea/extensions/pangea_room_extension.dart';
-import 'package:fluffychat/widgets/matrix.dart';
 import '../../../../utils/fluffy_share.dart';
 import '../../../../widgets/avatar.dart';
