+++ lib/pangea/utils/archive_space.dart
@@ -1,4 +1,4 @@
-import 'package:fluffychat/pangea/extensions/pangea_room_extension.dart';
+import 'package:fluffychat/pangea/extensions/pangea_room_extension/pangea_room_extension.dart';
 import 'package:fluffychat/pangea/utils/error_handler.dart';
 import 'package:matrix/matrix.dart';
